fix struktur folder, benerin all function admin & users

+++ admin/deleteUsers.php
@@ -1,5 +1,9 @@
 <?php 
 include 'env.php';
+if (!isset($_SESSION['username']) || $_SESSION['username'] !== 'admin') {
+    header("location: ../login.php");
+    exit();
+}
 if (isset($_GET['username'])) {
     $username = $_GET['username'];
     $query = "DELETE FROM users WHERE username = '$username'";
